+++ pretext/constants.py
@@ -141,6 +141,7 @@ ASSET_TO_XPATH = {
     "mermaid": ".//mermaid",
     "myopenmath": ".//myopenmath",
     "dynamic-subs": ".//statement[.//fillin and ancestor::exercise/evaluation]",
+    "references": ".//biblio"
 }
 ASSETS = ["ALL"] + list(ASSET_TO_XPATH.keys())