+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Instructions: Add a subsection under `[Unreleased]` for additions, fixes, change
 
 - New "handout" division, similar to worksheets, but intended for pages you print without exercises.
 - `@workspace` can go on most paragraph-level elements in worksheets and handouts.
+- Experimental support for improve bibliographies using citeproc.  `pretext generate references` will generate the required xml file in generated assets.
 
 ### Fixed